Refactor: user-story module 분리
